backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
groq
chromadb
python-dotenv
markdown
